instructor + student + grade script
Created the master grading script that combines instructor files (e.g. headers and a makefile) with student-provided files (e.g. implementation cpp or python) and produces output about how many points were earned.

The current commit does not work as-is, but the files are all present and (once the patching issues are sorted) correct.

Signed-off-by: Gabe Johnson <gabe.johnson@gmail.com>