Debugging module load/persistence issue
I've been chasing a bug caused by my badly-designed strategy of having 'Assignment.py' files that are loaded dynamically. This is a problem because the definition of Assignment sticks around in memory.

The fix is to use a flat text file (json, probably) and read values that way. This makes a bit more sense anyway, since right now Assignment subclasses AssignmentBase, but does not add any new functionality aside from defining job data like language, which files to use, and the name/point-value of tests.

Signed-off-by: Gabe Johnson <gabe.johnson@gmail.com>